credit_system/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
django==5.2.4
djangorestframework==3.16.0
et-xmlfile==2.0.0
kombu==5.5.4
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
prompt-toolkit==3.0.51
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
